fix(agent-runtime): move AgentCardProperties to engine.a2a package
AgentCardProperties was incorrectly placed in runtime.boot. Move to
engine.a2a since it is an A2A protocol concern, not a boot wiring concern.
Make hasExplicitName() public since it is now accessed cross-package.
